delete detail view controller and combine to root view
