+++ Address-Finder/Address-Finder/Presenters/AddressSearchPresenter.swift
@@ -7,6 +7,7 @@
 
 import Foundation
 import Alamofire
+import NMapsMap
 
 protocol SearchAddressPresenterDelegate: AnyObject {
   func presentAddress(result: [Documents])
@@ -14,7 +15,9 @@ protocol SearchAddressPresenterDelegate: AnyObject {
 
 class SearchAddressPresenter {
   weak var delegate: SearchAddressPresenterDelegate?
-  
+
+  //var selectedLocation = SelectedLocation(latitude: 37.345455350524844, longitude: 126.68751058508818)
+
   func doSearchAddress(keyword: String) {
     let headers: HTTPHeaders = [ "Authorization": "KakaoAK 754d4ea04671ab9d7e2add279d718b0e" ]
     let parameters: [String: Any] = [ "query": keyword ]
@@ -42,6 +45,23 @@ class SearchAddressPresenter {
   
   public func setViewDelegate(delegate: SearchAddressPresenterDelegate & UIViewController) {
     self.delegate = delegate
+  }
+
+  func setCamera(mapView: NMFMapView, selectedLocation: SelectedLocation) {
+    let camPosition = NMGLatLng(lat: selectedLocation.latitude, lng: selectedLocation.longitude)
+    let cameraUpdate = NMFCameraUpdate(scrollTo: camPosition)
+    mapView.moveCamera(cameraUpdate)
+  }
+
+  func setMarker(mapView: NMFMapView, selectedLocation: SelectedLocation) {
+    let marker = NMFMarker()
+
+    marker.position = NMGLatLng(lat: selectedLocation.latitude, lng: selectedLocation.longitude)
+    marker.iconImage = NMF_MARKER_IMAGE_BLACK
+    marker.iconTintColor = UIColor.red
+    marker.width = 25
+    marker.height = 30
+    marker.mapView = mapView
   }
 }
